LPS-77775 Inline HttpServletRequestMockImpl
